Fixed: if a copy of a  'Tree' object is created, as 'TreeTools' does on several occasions, the 'combiner' function is missing from the copy
Changes were made to make sure the 'combiner' function is included in copies of the 'Tree' object and the 'combiner' function is not considered if a  'Tree' object is compared
That is, because multiple function values which are created from a method are not equal.  This is a common scenario when a tree of pagelets is declared.

add some 'TreeTools' debug logs